requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
install==1.3.5
pillow==10.2.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
